+++ drf_proj/apps/client_api/v1/CHANGELOG.rst
@@ -1,8 +1,16 @@
 API Changelog
 -------------
 
+v1.1.0: changes
+===============
+
+- 2016-08-01
+    - `/api/login`: now accepts `email` instead of `username`
+    - `/api/test`: added a new endpoint
+
+
 v1.0.0: changes
 ===============
 
 - 2016-08-01
-    - added auth api: `/api/login`, `/api/logout`
+    - added auth API endpoints: `/api/login`, `/api/logout`